need to add future builder to appliances page
